Remove RedirectsAndIncomingLinks multi-cluster support
This class is only invoked with a constant connection, so there is no
need to support varied connection input. Additionally now that this is
no longer called via hook move away from global state and have Updater
directly instantiate. The awkward method of integrating the building
methods will be addressed in a followup patch.

Change-Id: I14de416ff412dfaa3a64322ac11ce4946f2873f3