Split up primary document building
To split document building with some parts happening
prior to the ElasticaWrite, but with bulk content loaded
within the ElasticaWrite, we need a clean api against which
to build documents.

Define an interface similar to RedirectsAndIncomingLinks
for building docs, and introduce BuildDocument class to
orchestrate building docs and applying implementations.

Change-Id: Ifa2469bcc1966ccffb35d1e7d5ad2a5b25929af6