Move bulk content for update after ElasticaWrite
Creates a new method, finalize, on the BuildDocument /
PagePropertyBuilder abstraction. This second stage of building is
invoked from DataSender::send, essentially on every attempt to write
to elastic. Adjust ParserOutputPageProperties to perform it's data
loading in the finalize stage, with the other builders remaining
constant.

The primary purpose here is to not write the bulk content, potentially
megabytes, to the job queue on each retry.

Bug: T236186
Change-Id: Ia0a079e2598f09e7c46ca19607875985bdf67fd6