Restore CirrusSearchBuildDocumentParse hook
It looks like I missed WikibaseMediaInfo when reviewing where this
hook was used. We stopped indexing structured metadata on commons
on oct 30 due to the hook removal being deployed. After deploying this
hotfix we will need to reindex updates to commonswiki for the last
two weeks, and think about how to fit wbmi into the content handler
interfaces for defining fields.

Bug: T237849
Change-Id: I79202fd5e4ac2f7fe5b13ab24ba4b0197484f131