Don't pass arrays into Status i18n messages
While typical i18n usage is to wrap numbers, the i18n exposed through
the Status object does not support this. It instead ends up converting
the array to string directly inside Status::getErrorMessage. It is
much too late to change the Status object interface, so pass bare
numbers which it accepts.

Bug: T237559
Change-Id: I052e3b029707d91583f6ac968ac42c4b17067485