Prefer strict assertSame() over lose assertEquals()
This is relevant for code that might either return an int or float.
It's even more critical when the expected value is 0. assertEquals() is
so loose, it will accept null, false, and all other values PHP considers
"falsy", even empty strings. A method that is expected to return 0 should
not return an empty string without a failing test.

A bunch of other edge-cases like this exist, mostly related to "falsy"
and "truthy" values.

Change-Id: I6b1a20a4d0a4ad36fb9fcabab0eab4a49eaf9b10