Apply FallbackMethod changes from atomic units
For our frontend analytics code to accept and report metrics about the
fallback methods that were run we need to track the provenance of the
query suggestion. Accomplish this by re-working FallbackMethod::rewrite
to have a concrete set of possible actions and using those actions to
track provenance.

Implements a new object, FallbackStatus, which represents the atomic
change to be applied to the search result (such as suggest a query,
add interwiki results, etc.). With this abstraction we track the name of the
fallback configuration along with action to be applied. This trace is
then simplified for output and provenance is reported to the frontend
via a javascript config named `wgCirrusSearchFallback`.

Bug: T238246
Change-Id: I69dbe185d8ed140fda2291ce6713d4be2e64d52e