Handle query suggestion snippets with HtmlArmor
Force callers to explicitly mark their snippets as html safe when
providing search suggestion snippets. Any plain string provided
will now be encoded, making it much harder to accidentaly send
un-escaped strings through.

It might be nice if this could be extended down through mw core with
updated search apis, but that is out of scope for this patch.

Depends-On: I76231d6fc53c4982eb4cd174d2e6a75eb2740497
Change-Id: I0275916cbbbb2ee9bb64c92192529991f0cf53db