Don't log with reserved parameter
Structured logging accepts an array of parameters to store, but some
of these are reserved and should never be used by an application.
Specifically the `message` key should not be set. This parameter doesn't
go into our bulk logging schema (CSRQ), only to the standard mediawiki
logging, so nothing special needs to be done to rename the parameter.

Bug: T245303
Change-Id: I6a57ee1d5293734a43417c8398c014fe2afac269