Wait around after a refresh before counting docs
While running a cluster wide reindex we are seeing around 2% of
attempts fail due to count mismatches. Inspecting these indices in
the following days shows no such mismatch in counts. We could
potentially drop this code all together and trust the elasticsearch
reindex API to work (this code predates the reindex api when we did it
ourselves), but as a first less-drastic step lets try waiting around
after the refresh completes to ensure it has finished. The refresh
here is particularly important because we disable automatic refreshes
while the index is being populated.

Change-Id: I2e7b0031aba8e6f724bcc6337a787925a5081c90