reindex: Wait around for counts to match before giving up
When running the in-place reindex procedure across all wikis some
(~2%) of wikis are failing due to counts, but the counts are later
correct. Wait around even more for the elasticsearch refresh to take
hold, even though per the docs when the refresh api call returns the
refresh should be complete.

Bug: T247368
Change-Id: If0d495aa74ce936fc535dfa627c9bcd476deb6ac