Only run certain big, isolated tests in this repo, not in gate
This means that patches in other repos will run CirrusSearch's tests but
exclude these tests from the suite. This is intended for tests which are
isolated from the rest of the MediaWiki system such that changes made in
MediaWiki itself or other extensions won't change the behaviour of those
CirrusSearch components and systems, and slow enough that it's worth the
effort of singling them out.

In practical terms, this moves 10 of the 16 slow tests out of the 'unit'
suite run, saving about 4 seconds per run out of 10, and of phpunit's 50
slowest tests from the integration suite, saving about 60 seconds out of
203. This is a very substantial saving for all patches, and I'd strongly
encourage this to land if the team is OK with it.

I picked on the slowest unit tests in production extensions, and the bit
left after major surgery was performed on the SearcherTests, to only run
a random selection of 200 of them, in ad4c23c; that can now be reverted,
if the Search Platform team wishes.

As part of this, unit tests that are in @Standalone are moved to another
directory, so that they aren't run twice.

Bug: T221434
Bug: T225184
Change-Id: I2b31aaaf459e7e7e3d74f539d8e173055b62db42