Explicitly remove temporary hooks in FunctionScoreChainTest
MediaWikiIntegrationTestCase::setTemporaryHook() used to allow calling
setTemporaryHook() multiple times when setting handlers for the same hook,
as it would replace any existing handlers in $wgHooks[$hook] with the new
handler value it received.
To support the new MediaWikiIntegrationTestCase::setTemporaryHook(), which
no longer store temporary hooks in $wgHooks, we must
explicitly remove the hook before we can set another handler for it again.

Bug: T250300
Depends-On: I0c41f50a6416fd2fe5607278e31c0a07f28dd3d8
Change-Id: I8ca352e9b4679ea85154c50e04af86694ff5af29