Include concrete namespace mapping in config dump
External services can provide additional data to the search indexes,
such as popularity information or ML predictions. In the past they
have only written to the content index, allowing pages not stored
in the content index to fail. There are new use cases to update
files and drafts, to allow the external processes to send their
updates to the correct index provide a concrete mapping from all
known namespaces to the index the pages are stored in.

Bug: T250238
Change-Id: I9acb34c34a116c8502feae2981213f29dc9df136