InterleavedResultSet should implement SearchMetricsProvider
This looks to have been lost in a refactor of ResultSet in
I0e807425. The functionality was not removed, it looks like the
only thing necessary is to re-tag the class with the metrics
provider interface.

Change-Id: Ia4a4126179b53896b55fc38869eecd73216dc18b
(cherry picked from commit 3b1115cb24fba2c16cabb3b541998ced2b15dc3b)