Simplify creating indices on all writable clusters
Even though cirrus knows what clusters it can write to, there was
no way to ask it to initialize a new wiki on all clusters. Previously
the addWiki.php script in WikimediaMaintenance was inspecting cirrus
config and running the script for each, but that should move inside
cirrus.

Adds an `all` cluster option to UpdateSearchIndexConfig along with
looping the child executions over the set of selected clusters.

Bug: T254331
Change-Id: I2e43ff29418f0db6344235a04ce152bee696b4c8