Add PageIdFeature
Adds a new pageid: search keyword which restricts the query to the
given page IDs. This is useful for re-validating cached search
results.

Also exempt pageid: from length check. Since it is used to check
cached page sets, it can have a large number of arguments, and
there is no risk of performance problems with straight document
ID checks.

Bug: T269642
Change-Id: Ic9eda738a3311ad781d26e65034f3e2fb35bcf4d