Use standard pipe separator for inlanguage: keyword
Almost everything in mediawiki uses the | as a separator, including
the other search keywords. Lets be consistent and use the pipe here
too.

In an analysis of queries to WMF wikis in january just shy of 40%
of queries invoking inlanguage also have a comma somewhere in the
query. Not sure when we can actually hard deprecate, but this is
the first step and we can check the stats at a later date.

Change-Id: I6903218750c4d5a83d366029938752772f4d475a