Make weighted_tags updatable from CirrusSearch using the multilist handler
Expose a function resetWeightedTags( $pageIdentity, $tagFamily ) from
CirrusSearch allowing to reset a tag family for a single document.
Add necessary hints to set the multilist handler on this field.
Properly cleanup the document in case of upsert if it contains __DELETE_GROUPING__
magic word to avoid adding it as the document data when inserting.

Bug: T273508
Change-Id: Ic9acbd38fbb92e62dc9ea6544bb50483b7e0ca30