Improve logging of update errors
List the error details for each error in case of bulk requests,
and traverse the error's caused_by tree to explain the error better.
E.g. where the current exception message is

    update: /wiki_content_first/page/56 caused failed to execute script

the new message would be

    update: /wiki_content_first/page/56 caused failed to execute script
      - update: /wiki_content_first/page/56 caused failed to execute ⏎
        script: illegal_argument_exception: failed to execute script ⏎
        (java.lang.String cannot be cast to java.util.List -> Expected ⏎
        a list, but recieved (java.lang.String) -> Failed updating ⏎
        document property weighted_tags)

which, while unwieldy, contains key information for understanding what's
going that was omitted from the previous version.

Change-Id: I1185467df2c7758b1821bde85853dff25cd2f916