Add helper method and script for updating weighted tags
Adds CirrusSearch::updateWeightedTags which works like
CirrusSearch::resetWeightedTags but replaces tags
instead of deleting them. Also adds UpdateWeightedTags.php
maintenance script to expose this functionality.

This is primarily meant for easier testing of software that
involves searching for tags. In production, a more efficient,
non-MediaWiki means of updating tags would be used.

The syntax for weights is a bit awkward, but there didn't seem
any better way, given that all tags with the same prefix have
to be set at the same time so they don't overwrite each other.

Bug: T272304
Bug: T261409
Change-Id: I02f311d245f54c494daf4e12002dfdf8c94d5f98