Add test for SpecificAliasValidator
elasticsearch/elasticsearch 6.8.0 was just released and this seems to
get picked over 6.7.2 which was the previously used version by
ruflin/elastica. However 6.8.0 has removed a large chunk of code that
gives errors as described in the ticket.
This adds a test that ends up calling the Index functions that was
removed.

Bug: T276320
Depends-On: If00e4f871dbfa5572563f1eb5f60ded51f79a06b
Change-Id: Ia89bba7da18ea722691a4f65758aa187d7d70130