Extend query length limit exemptions to negated keywords
Some keywords, such as hascategory, are allowed to have long
parameter lists which do not count against the query length limit,
but due to the way the length check works, this did not apply when
those keywords were negated. Since the performance implications
are be similar, and negative filtering for long lists can also be
useful, extend the exemption to such cases.

Bug: T276827
Change-Id: I4cfff0bd3e682111623d1c6b801356da1328e84c