Reconcile configured indices with live state
CirrusSearch multi-cluster configuration allows spreading thousands of
indices across multiple clusters, and can maintain multiple replicas
of the same index. In a complex deployment this can result in many
thousands of indices, with no simple way to verify the set of indices
that exist are strictly the required set. Add a process that
reconciles cirrussearch configuration against a cluster for a set of
wikis.

The reconciliation process is accept-list based. Some number of index
acceptors can be configured per-cluster. All indices must be claimed,
unclaimed indices are reported and attempts are made to explain how
that index came to be.

No attempts to automatically resolve problems is currently provided,
but could be added in the future as we gain confidence in the output.

* Add a maintenance script to report the set of indices cirrus is
  configured to use across all configured clusters.

* Add a python script that aggregates the maintenance script output
  across a fleet of wikis and compares it against the state of multiple
  elasticsearch clusters. This is fairly tied to specifics of the WMF
  deployment, but should be generic enough to be reusable elsewhere
  with extraction of the stage that builds the configuration.

Bug: T279607
Bug: T279009
Change-Id: Idc07fae6f13caaa0af048f2ab1ceec8626c6dbb9