Limit load generated by Reindexer auto-slicing
Attempting to reindex large indices on cloudelastic is causing
more load than the cluster can take. One potential issue is
CirrusSearch is auto-slicing to 30 tasks, but the cluster only
has 6 nodes.

Reviewed elasticsearch docs and code, in 7.x they added their
own auto-slicing option which caps at 20 slices. I've added
that cap to our own detection and added a second limit on the
number of nodes in the cluster. This is only changing the default
behaviour, operators can still provide an explicit value on the
command line.

It's not clear what the full purpose of the limit of 20 tasks is. Our
current performance is less than stellar though, so trying out
what upstream recommends seems reasonable.

Bug: T280601
Change-Id: I47fc1cc0b41741745c9bfab1653e9d6e520cd42f