Refactor UserTesting to assign test buckets
Previously test buckets had to be assigned by a separate javascript
component and provided as part of the query string. This had the
downside of not being able to assign test buckets to requests starting
outside the typical UI of the site, for example through a browser
search bar or link. A large cohort only issue one or two search
requests, so this ends up excluding them from the measurement.

Implement a method which automatically assigns users to a test
bucket for all index.php requests, and can still be manually triggered
for api.php, rest.php, etc. Further documentation included in patch.

Needs testing and frontend component (in different repo) to further
verify operation. Needs review of the docs, certainly still
inconsistencies.

* Moved documentation of testing from docblock of UserTesting
 class to docs/ and updated to match the new reality. Doc primarily
 covers how this is intended to be used.

* Split UserTesting into UserTestingEngine and UserTestingStatus.
 UserTestingEngine keeps most of the implementation, UserTestingStatus
 reports the result of a bucketing decision by the engine. Intent
 is to allow easily accessing decision without side effects.

* Add reporting of the autoenrolled bucket in cirrus-config-dump api
  call. Allows autocomplete sessions to source this information.

* Review and adjust the few places that directly call user testing
 code to use the refactored classes.

Bug: T262612
Change-Id: Iea49188d9a97feddd4ec1ff69235dbc1e6f7923b