Include tracing info in failing assert
We saw this assert fail a few times over a short timespan on a single
wiki. It's not clear what exactly went wrong, plausible the doc was
deleted shortly after the revision that triggered an indexing request
but we don't have the information to know for sure. Add the docId to
the assert fail message so if it happens again we have more info to
track down what exactly happened.

Bug: T282036
Change-Id: I577af3c048ba54c244b44b2c4a8b071e845470eb