Add pool counter for automated requests
The new pool is intended to limit requests identified as external
automation.  We've had issues both in the past and currently ongoing
with automation from public clouds overloading the pools and our only
mitigation is to block their traffic outright. This pool gives a place
to direct that traffic where it can continue but in a controlled
manner with less impact to our interactive users.

Automation check is implemented as matching either a configured regexp
against the http user agent, or matching a configured list of subnets
(via array of CIDR strings) against the users IP address. This pool
counter for automation will override any selection other than regex.
Regex is particularly expensive and we wouldn't want to let both pools
fill with regex queries. This also doesn't apply to completion search,
which has a constant pool defined elsewhere, but we haven't really had
issues with that pool saturating.

Dependent patch must be merged first to ensure there is a configured
pool when the patch is deployed.

Bug: T284479
Depends-On: I3e8da19b2a51e469bf2c2d8367a87a7226ce17b4
Change-Id: Icfb115468c33f34fce1270127f69fbcc7bcf3741