Revert "Add support for ores drafttopic" and "Stop querying ores_articletopics"
This reverts commit 0a73dcb7214258f7abd06c0d27df2f3ed632049f
and c49990a6235d0d4e47430029089426ec9834788e.

Reason: We received reports that some wikis are no longer returning
results. In a review of rowiki that content index has not completed
a reindex yet, so the weighted_tags field does not yet exist. Revert
this removal of the BC handling so those wikis can limp along with
the old (and no-longer updated) values until we can review and
appropriatly reindex all affected indices.

The drafttopic keyword is being reverted at the same time out of an
abundance of caution, that patch builds on the one that removed BC
handling and would require resolving merges.

Bug: T285577
Change-Id: I68e1fefc1a72fe679e08abf0619ffd8b768e57b1