Provider a stable random sort
Provides a variation on the existing random sort to have the results
in a random order, but consistent across requests for a single user.
This is performed by seeding the random number generator for each
document with (weak user fingerprint, update sequence number). This
gives the same value to the same document, as long as the fingerprint
is still the same and the document has not been updated.

While the sequence number is the suggested default value to use, we
might have to revisit this depending on exact use cases. By using
the sequence number the value is cheap to lookup at runtime, but
it will change if the underlying document is updated. If specific
requirements arise we can easily revisit this.

Change-Id: Ic22ce802920054ea73047d126c00108f5aff4f59