Hack around i18n cache failure for wikitech
The problems preventing the completion index from building on
wikitech are somewhat systemic and will not be resolved in
a day or two. Plausibly we wont be able to address the root
cause until the separation of wikitech and horizon is complete,
allowing wikitech to become more like the rest of the wikis
instead of in it's own isolated place.

Bug: T288233
Change-Id: I00cdf582f9211293d4344ade75e9e323425597e1