Support for partial clearing of weighted_tags
When loading new data into weighted_tags, data in unreferenced
documents remains behind. To support clearing the old documents
provide a script that will, when given a list of wikis and an
indexed value, clear that prefix out of all matching documents.
To avoid clearing the new data as well a CSV can be provided listing
all page_id's to exclude from the clearing process.

Bug: T295316
Change-Id: I2d6577003d103b48423e414c1e5909ad63f4f607