Increase minimum integration test wait time
Seeing the 'To' page edit failing intermittently due to timeout. Make
a guess that the problem is other edits get to merge their timeouts
together papering over the ones that take too long, but this edit is
in a batch by itself and only received the 10s timeout. Increase the
minimum timeout we will wait to 20s even if there is only one item.

Change-Id: I0ed7154aab3f31d95af57fbcf1e46b41ed6da60b