Re-add option to run ElasticaWrite in-process
We initially used the job queue to isolate write jobs from each other
when we had two full size clusters and a third smaller one that would
fall behind. While it has worked for some time, lately the job queue
has not been able to keep up with the rate of jobs we are creating.
Add an option that will allows us to only isolate the small cluster
which should hopefully increase the rate at which we can apply
updates, and allow us to re-enable the saneitizer.

Change-Id: I6ef39c67ee614a165df68995ae3d80db41251624