Convert empty result set from anon to concrete class
BaseCirrusSearchResultSet::emptyResultSet() returns an anonymous class, which
throws on serialization attempts. That will be included in the Status returned
by SearchEngine::searchText() and similar.

There is no requirement for Status objects to be serializable (all the
documentation says on the matter is "The use of Message objects should be
avoided when serializability is needed.") so this is not exactly a bug, but it
would be easy to improve by turning the empty result set into a proper class.

Bug: T302473
Change-Id: Ic232d76426e12934685baf9819a82715229f34e7