Remove freeze/thaw index support
This functionality was added when we were using elasticsearch 1.x and
restarting an instance required waiting 40-120 minutes for the cluster
to recover all the shards into a healthy state. Elasticsearch has
improved significantly since then, an instance can be restarted
without freezing the cluster and the cluster will report healthy
within a couple minutes.

Given the current state of elasticsearch this complication doesn't
seem worthwhile. Additional followups will be required to remove this
concept from our SRE side as well.

Change-Id: Iabc53bbbfe3afe5fe1b9230232573a0c1fd11be2