tests: skip namespace lookup in request logger test
These fixtures end up being tedious to rebuild, as they make
assumptions about the state of elasticsearch that are partially
provided by running the browser integration suite, but also depend
on specific page id's existing and not existing.

To avoid rebuilding the fixtures for a version check being added to
MetaStore configure the test to not query namespaces through
elasticsearch. This is also the suggested configuration for
deployments, the only reason it isn't the default is that not all
instalations will have the php-intl extension we require.

Change-Id: I9b03763065d6c271ce30f665bb4c649589f8ec90