include_type_name: Index creation and status
When creating or checking on indices we trigger a variety of
include_type_name deprecation warnings. Be explicit about where we
expect the type names to be included, and where we do not. For the
most part, for now, ask the names to be included rather than doing
much refactoring. These can be revisited once we are on 7.x

Bug: T288764
Change-Id: I80b8cb909d96d9574b0a4ac920c4ce2a759bbd20