Prevent pathological wildcard queries to reach the backend
These queries were able to bring the cluster down in the past and we
had to fix an issue upstream (T102589). Sadly a regression appreared
in the elastic7 branch causing this fix to be ineffective.
Try to detect such queries from CirrusSearch and don't send them
to the backend.
We might consider fixing this upstream.

Bug: T307931
Change-Id: I77c9dcf7d980590c01783328d387ff420b2d2308