metastore: Update deprecated version parameters
These parameters are having the leading underscore removed in 7.x.
While this will fix itself without a code change when we transition to
the 7.x version of the Elastica library, we also need the 6.x version
of cirrus to issue writes that are fully compatible with 7.x to
support both during deployment.

Bug: T218994
Change-Id: Idac975b403aeeefacd949b59072386d981f1c94f