Add partial support for es6
Painful hack done with a transport wrapper that rewrites bulk requests
to attach the mandatory "_type" to all operations.
Necessary to make this CirrusSearch branch able to write to ES6.

Usage:
$wgCirrusSearchServers = [
    [
        'transport' => [
            "type" => \CirrusSearch\Elastica\ES6CompatTransportWrapper::class,
            'wrapped_transport' => CirrusSearch\Elastica\DeprecationLoggedHttp::class
        ],
        "host" => "elasticsearch"
    ]
];

Added a couple hacks to make sure that the maint scripts don't fail on getting
the number of total hits.

Change-Id: I9999000d52ab1a9e620319b251dcbccf369b90ef