Only create archive indices when requested
Archive indices were still being created even if the configuration
turned them off. Adjust places where we check isPrivateCluster to
also check if archive indices are enabled.

This change exposed that the MetaVersionStoreTest could pass or fail
depending on host the host rnuning the test had been configured.
Adjusted to define the exact configuration to test under in that
case instead of importing the host configuration.

Change-Id: Ica00b2cb3b87d9b89eff1542d80891344b5552a5