Add ability to disable skip_if_last_matched
The hardcoded skip_if_last_matched behaviour for specific fields is
really an optimization for how the Special:Search UI is rendered,
trying to avoid overloading the user with multiple similar highlights.
But other use cases, such as API calls, need more flexibility. Add
an engine feature that allows toggling this behaviour off.

This is paired with a patch to mediawiki/core that provides the
'snippets' hint including a list of fields it will report snippets
for. Currently the exact list is ignored and we simply toggle the
behaviour on or off, as the overall performance impact is expected
to be minor. We can't simply turn this off everywhere as the
Special:Search UI doesn't gate snippet display on anything and
it would immediatly add additional noise to the results page.

Bug: T306477
Related: Ied664e3c1b7d246bca914f10c8fd2e36ff6418ec
Change-Id: I38d58a5af45eef7cd1798312af137a70f73c54c3