Truncate failed requests errors to 4kB
Failed writes in WMF prod don't end up in the normal logging places,
because the logs generated are too big and get truncated. The logging
pipelines look to be limited to 32kB, but this logs multiple MB at
times. Truncate down to 4kB to get an idea of whats wrong without
blowing out limits.

Change-Id: Ib44646cea7efdd998ae9bea918def374225c16f3