metastore: Remove versioning from saneitize updates
By default elastica is loading the version field from elasticsearch
and then providing that version back to elasticsearch on update. This
complicates the elastic 6->7 transition as we would need to move over
to the `if_seq_no` and `if_primary_term` parameters.

In I69926b4b we decided to drop the versioning all together, but the
removal was incomplete as we didn't notice that elastica provides
these by default. Adjust the updates to remove the versioning
parameters that we weren't exepecting.

This isn't a problem for the normal update pipeline because those
updates are all scripted through super_detect_noop and don't use
follow the same code paths.

Related: I69926b4b7d3dbd719fd429c0d330ab2c33331854
Change-Id: Iafd2cf9631eb678e0a248f251b6ac3e018d9cd0b