Search results: add extra fields conf
For now we have in code EXTRA_FIELDS_TO_EXTRACT feature so we can get in results (fields in response of query).
But there is no way to enable it from default search API.
We added config for that, $wgCirrusSearchExtraFieldsInSearchResults.
When populate with an array of strings, the query would grab those fields from document match to the search
Those fields must be in the index, by $wgCirrusSearchCustomPageFields or by hook

Bug: T311550
Change-Id: I8ed095e48e92fc9a57c2c1559456fbb4fb81a23c