reindex: Detect index type from live mappings
Reindexing from a `page` type to a `_doc` type currently fails as it
attempts to write to the `page` type anyways. Providing _doc on both
sides equally fails, it seems elasticsearch's aliasing of _doc to
the live type is incomplete (also seen elsewhere).

Resolve by querying the index and using whatever type name is
currently defined in the mapping. Once everything has been reindexed
to use _doc this can be removed.

Change-Id: I3b177cb46adaa796b883d64f059307f6aa230dba