Add explicit partitioning key to ElasticaWrite
Adds a key to ElasticaWrite jobs that a jobqueue implementation can
use, if so configured, to place the jobs for the same cluster into
independant partitions. While plausible that a job queue could split
a cluster into multiple partitions itself the targeted queue,
cpjobqueue, can only currently do static assignments from a
partitioning key to a partition number. To support this provide
multiple values for a single cluster that it can use to partition
the resulting queues.

Bug: T314426
Change-Id: I1496b915c7a90a493af3c11d4bafb96f4cc3a844