Ensure namespace filters is passed as a list
When the namespace array is mutated, such as slicing out NS_MAIN to
perform a comp_suggest query, the resulting array looks like an object
to php's json encoding, which elasticsearch interprets as being the
start of a terms lookup definition. Take an array_values of the
namespaces to ensure it encodes into a list when sent to
elasticsearch.

Change-Id: Idec04cb5caeb755009cbab517c750c2a46ddad4a