Add token_count subfield to outgoing_link
This allows the use of the number of outgoing_link elements as a
scoring signal when performing search queries. This wont be used in
cirrus, at least not initially, but will be used within the customized
queries issued by add-a-link.

Not 100% happy with where this fits into the mapping config building
process, but it's not clear where else it would fit. The outgoing_link
field and it's associated flags come from mw core. The field flags
themselves are a bitfield, it's not really safe for cirrus to add
custom flags that could later conflict with core flags. Additionally
we would need additional context about what analyzer should be used
when counting.  Avoid the complexities and simply add the subfield
directly.

Bug: T317546
Depends-On: I53219679af139b1b9385dad130449e049456c16e
Change-Id: Ie42cf38cf55daaf2609f21ce8fbe7c7a413b423f
(cherry picked from commit 7ebfb2269afc8df56682077ec5fd1f2e6bad4dd6)