IndexCreator: Wait for green after creation
We are seeing some tedious errors during reindex against wikibase
instances, they often report that the index has finished creation
but then 404 when we try to validate the index settings. Review of
cluster state after the failure shows the index was created, it seems
like we are hitting a race where the followup arrives too quickly.

Instead wait around for the index to reach a green state before
continuing. This might take a bit longer but should make reindexing
less error prone. This makes the output slightly more awkward,
reporting the waitForGreen status during index creation, but we want
to ensure if exceptions are thrown the messages get printed out.

Bug: T317200
Change-Id: I14fd7b3eed7c15af86cc0fddcbe025e4b5adb1c6