Remove support for elasticsearch 6
With mw 1.40 branched and the compatability code available there we
should be safe to remove elasticsearch 6 compatability from
CirrusSearch.

To create this patch I reviewed the patches merged to the es710 and
es68 branches, prior to merging them into master. This is likely not
every last bit of compatability code, but it should be most of it.

* Removes various include_type_name => false parameters, this is now
 the default.
* Remove Util::getIndexMapping, with es7 the default we can use
 Elastica\Index::getMapping()
* Remove support for creating indices in es6 from ConfigUtils
* Remove ES6CompatTransportWrapper, it was only needed while talking
 to mixed clusters.
* Remove UpdateSuggesterIndex::getTotalHits, with es7 the default we
 can use Elastica\ResultSet::getTotalHits again.

Change-Id: I74851aff38a146c26c8e4cd85ff27b6685fc5db1