Add DocumentSizeLimiter a component to limit cirrus doc sizes
DocumentSizeLimiter is a class that attempts to limit the size
of a CirrusSearch document by applying few heuristics.
They heuristics can be defined in a profile and a set of profiles
(based on WMF experience) is provided out-of-the-box.
- default: which does nothing
- wmf: which limits the size of file_text and opening_text to 51200
  and 10000 respectively
- wmf_capped: applies the same limits of wmf but will apply more
  if the document size is above 4Mb. A set of fields will be truncated
  up to a minimum limit obtained from 99.999 percentiles of the field
  size of WMF hosted pages. The template CirrusSearchOversizeDocument
  will be tagged to such docs.

The algorithm is approximative and there are no guarantee that the
expected limit will be honored. The document size is obtained by
serializing it to json.

Bug: T316016
Change-Id: I4b1505e7cd1a3f51f36715652bbd7d28189fa5ad