Add config option to disable incoming link counting
Counting of incoming links are quite frequent and induce a constant
query load on the search instances, but for the most part the results
of those queries are thrown away because the value hasn't changed more
than 20% from the value already stored in the index.

Add the ability to disable those queries when an external process is
able to update them. In the WMF case an aggregation is performed
over dump of the search indexes to count the incoming links and issue
updates only to pages that have had their values changed.

Bug: T317023
Change-Id: Ib634d17e188686a24657bb41e2424bf45b46d6ba