SearchAfter: Increase default retries to 12
The prior retry setting of 5 only gave 15s for things to recover
before failing the operation. While this is enough for some cases,
we've seen dumps against large wikis (wikidatawiki) fail after a node
reboot because subsequent requests weren't able to run in time.

Increasing the retry count to 12 allows the backoff to scale to it's
max value, 120s, and gives a total of 607s (10m) for operations to
recover before the retries are declared a failure.

Bug: T265056
Change-Id: I49632c8afe13499bacabdccded8f170ed7dcb6b4