looksLikeAutomation: Allow flagging requests from arbitrary headers
This was limited to only recognizing if a request was automation based
on the http header, but there are other headers that might be viable.
Particularly in the WMF context there is upstream infrastructure that
flags requests as coming from public clouds (aws, gcp, linode, etc.)

Generalize the support around flagging requests based on the
user-agent http header to generically work on any http header.

Bug: T326757
Change-Id: Ifeca2410507d459d90fe46780668e22f2b8e1be0