Add WANCache to ParserOutputPageProperties::finalize
According to flamegraph, every call to write to the cloud basically
doubly parses the same page it was rendered before:
https://performance.wikimedia.org/arclamp/svgs/daily/2023-05-11.excimer.RunSingleJob.svgz

This should easily brush off ten percent of resources from job
runners.

It's not the cleanest solution, I understand. But my hope is to get this
deployed and measure the impact and then improve or revert.

Bug: T336698
Change-Id: I24a5431149be84fc95618f902b6bb758d1ba7460