Add a config flag to enable analysis chain deduplication
We've been having issues with WikibaseCirrusSearch indices for some
time. They are slow to create, can trigger master_timeout issues, and
generally don't behave like the other indices. We've tracked this down
to an issue in elasticsearch with how it handles large amounts of
index settings. By dramatically reducing the number of settings
through deduplication we can get the behaviour of these WBCS indices
mostly inline with the normal indices.

This patch also allows the dead analysis chain detection to run for
all contexts. This is a prerequesite to the deduplication but was
previously only enabled on archive indices out of an abundance of
caution, and it seems to have run well in that capacity.

Bug: T334194
Change-Id: I3f70878126f4429a8a3aa21eab020bd6cd80af8f