Attempt to properly handle queries with broken boolean logic
Inspect errors generated from the QueryStringRegexParser and immediately
fallback to a degraded query.
Fix buildDegraded which was broken since it did not replace the
highlight query that still contained the query_string query.
Disable all costly operators on the simple_query_string as it seems
it was not run for a while.

Change-Id: I518a96b70e8c5233fd45ff5c8d302eec23551e60