Add very basic integration tests for wikibasecirrussearch
Arguably these tests should be in the wbcs repo, but it's not
currently simple to spin up a small integration suite there.
For now, add some simple tests that verify the very basics of
wbcs functionality works, namely that an entity can be found
by label it both autocomplete and fulltext search.

While here, simplify the way we source the bot password for
integration tests. We require defining a separate env var
for each wiki, but in practice we have a single value used on
all wikis. Make them all read from the same env var to better
support the current use case.

Bug: T339935
Change-Id: I87ffcbdd4c1f2f8a75a600e4d9f9d1e4298279bf