Make tests that use ExtensionRegistry integration tests
All these tests end up calling ExtensionRegistry::getInstance(), which
is unsafe in unit tests because it adds a dependency on all installed
extensions.

ExtensionRegistry should be injected where needed, but that seems
hard-ish to do because it's used in a lot of places. For instance,
injecting into ConfigProfileRepository would already be pretty hard
because the class is instantied a lot of times.

Make the tests integration tests instead, as a temporary measure until
ExtensionRegistry is injected.

Needed-by: I0a04c82250582fed7a66c1e10868d9b4f3823a28
Change-Id: I4e678e996845e140aef17724eda3e13b5b0eea08