Try to identify page changes the same way EventBus does
The states provided by LinksUpdate do not seem to provide enough
information for us to determine what are the links update that are
triggered in the context of page change.
This patch attempts to identify those by listening to the same hooks
that EventBus PageChangeHooks listens to (process that is responsible
for producing the mediawiki.page_change) and keeps track of a state
that can be used when the LinksUpdate hook fires.
This process is definitely fragile as it introduces a temporal
dependency that we try mitigate by deferring the actual run of the
ChangeListener::onLinksUpdate process.

Bug: T325565
Change-Id: Id9056528a433faf0b3260d8ac9d2136dab068627